+++ .gitignore
@@ -9,3 +9,4 @@ outputs
 # Log files for colab-convert
 cc-outputs.log
 *.mp4
+scripts/deforum_helpers/src/rife/inference_video - latestnotworkingpy
