+++ scripts/deforum_helpers/frame_interpolation.py
@@ -1,0 +1,5 @@
+from rife.inference_video import *
+
+args_dict = {'video': '20230110202008.mp4', 'multi': 2, 'model_name':'RIFE43'}
+#args = parse_args(args_dict)
+parse_args(**args_dict)
